WED fixes part 12

+++ Sources/WorldEditor/Tree/base_property_tree_item.cpp
@@ -58,7 +58,7 @@ BasePropertyTreeItem* BasePropertyTreeItem::child(int row)
 
 int BasePropertyTreeItem::childCount() const
 {
-  return m_childItems.size();
+  return static_cast<int>(m_childItems.size());
 }
 
 BasePropertyTreeItem* BasePropertyTreeItem::parentItem()
